% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\name{aw_exists}
\alias{aw_exists}
\title{aw_exists}
\usage{
aw_exists(entity = NULL, id = 0)
}
\arguments{
\item{entity}{Ein Entitäts-Typ als String (vgl.: entities)}

\item{id}{Die aw-ID der gesuchten Entität}
}
\value{
TRUE, wenn die Entität mit dieser ID existiert, sonst FALSE
}
\description{
Hilfsfunktion prüft, ob Entität mit dieser ID existiert
}
