% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\name{aw_get_id}
\alias{aw_get_id}
\title{Datenabfrage}
\usage{
aw_get_id(entity, id, ...)
}
\arguments{
\item{entity}{Entity class, as listed in the factor entities}

\item{id}{}

\item{...}{List of filters}
}
\description{
`aw_get_id` fragt Daten für ein Objekt nach seiner ID ab
}
\details{
Abfrage eines einzelnen Objekts aus der aw-Datenbank. Gibt ein Listen-Objekt zurück, das dann mit der üblichen R-Notation mit $ ausgelesen werden kann. Der Typ der Entität muss bekannt sein und angegeben werden; IDs sind nicht eindeutig: 9188 bezeichnet ein Entität vom Typ candidacy-mandate und vom Typ constituency.
}
\examples{
aw_get_id("politicians",1)

}
