% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\name{aw_get_table}
\alias{aw_get_table}
\title{Request table of objects}
\usage{
aw_get_table(entity, ...)
}
\arguments{
\item{entity}{Entity class, as listed in the factor entities}

\item{...}{List of filters}
}
\description{
`aw_get_table` requests data as a data frame.
}
\details{
This function is used to request one single object, belonging to an entity class,
defined by its ID.
}
\examples{
aw_get_table("parliaments")

}
