% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\name{aw_kandidaten}
\alias{aw_kandidaten}
\title{aw_kandidaten}
\usage{
aw_kandidaten(id = NULL)
}
\arguments{
\item{electionID}{- aw-ID der Wahl bzw. Wahlperiode}
}
\value{
df mit ID, Namen, Partei-ID, Partei, Wahlkreis, Wahllisten-ID, Listen-Position, Wahlkreis-ID, Wahlkreis-Name
}
\description{
`aw_kandidaten` gibt eine Personenliste für eine Wahl- oder Wahlkreis-ID zurück.
}
\details{
Diese Funktion prüft, ob sie mit einer gültigen Wahlnummer aufgerufen wurde; wenn ja, gibt sie alle Kandidat:innen für diese Wahl zurück. Sonst nimmt sie an, dass sie mit einem Wahlkreis aufgerufen wurde - Wahlkreis-IDs sind ja immer fest einer Wahl zugeordnet - und ruft alle Kandidaten ab. 
Wenn sie mit einer Wahlkreis-Liste aufgerufen wird, zieht sie sich die IDs aus der Liste.
}
\examples{
aw_kandidaten(aw_wahlkreise(128)) alle BTW-Kandidat:innen 2021
aw_kandidaten(aw_wahl("Bund",2017)) alle 
aw_kandidaten(10225) alle Bewerber:innen aus dem BTW-Wahlkreis Kassel 2021

}
