% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\name{aw_wahl}
\alias{aw_wahl}
\title{aw_wahl}
\usage{
aw_wahl(p, y = year(today()), strict = FALSE)
}
\arguments{
\item{p}{ein String mit dem Namen des Parlaments ("Hessen" - auch in der Form: BLxx mit der Länderkennzahl aus der AGS, etwa "BL06" für den Hessischen Landtag); wenn p numerisch ist, wird es als aw-ID des Parlaments identifiziert (z.B.: 11 für Hessen)}

\item{y}{eine Jahreszahl der gesuchten Wahl bzw. Wahlperiode}

\item{strict}{Boolean-Variable: TRUE, wenn nur genaue Treffer akzeptiert werden sollen; in der Default-Einstellung sucht die Funktion nach dem nächsten passenden Wert}
}
\value{
Integer-Wert mit erstem Treffer
}
\description{
`aw_wahl` gibt die ID(s) für eine Wahl zu einem Parlament zurück
}
\details{
Mit aw_wahl kann man sich die IDs einer Wahl suchen lassen, und zwar entweder über den Namen (oder die Länderkennzahl) oder die aw-ID, dazu das Jahr (es wird das nächste gültige gesucht), und die Einschränkun.
}
\examples{
aw_wahl("Bundestag") liefert die ID der aktuellen BTW
aw_wahl("Hessen 2021") liefert die ID der nächsten Landtagswahl 2022
aw_wahl("Bundestag",2021) liefert 128 (die ID der aktuellen Wahl)

}
