% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\name{aw_wahlkreise}
\alias{aw_wahlkreise}
\title{aw_wahlkreise}
\usage{
aw_wahlkreise(electionID = NULL)
}
\arguments{
\item{electionID}{- aw-ID der Wahl bzw. Wahlperiode}
}
\value{
df mit der constiuency-ID, der Wahlkreisnummer und den Wahlkreisnamen
}
\description{
`aw_wahlkreise` gibt eine Wahlkreisliste für eine Wahl-ID zurück.
}
\details{
Diese Funktion ruft zu einer Wahl bzw. Wahlperiode die zugehörigen Wahlkreise ab. Die IDs der Wahl bzw. der Wahlperiode ist immer fest einem Parlament und einer Zeit zugeordnet. Als Rückgabe bekommt man eine Liste der IDs, Wahlkreisnummern und Wahlkreisnamen zu dieser Wahl - die Wahlkreis-IDs sind ebenfalls fest einer Wahl zugeordnet, d.h.: ein und derselbe Wahlkreis (z.B.: Wiesbaden) hat bei der Bundestagswahl 2017 eine andere ID als bei der darauffolgenden Wahlperiode oder der Bundestagswahl 2021 - oder bei der Landtagswahl, die einem völlig anderen Parlament zugeordnet ist.
}
\examples{
aw_wahlkreise(128)
aw_wahlkreise(aw_wahl("Bund",2017)) oder
aw_wahl("Bund",2021) \%>\% aw_wahlkreise()
aw_wahlkreise(aw_wahl("Hessen",2018,wahl=TRUE))

}
