% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\name{aw_wahlperiode}
\alias{aw_wahlperiode}
\title{aw_wahlperiode}
\usage{
aw_wahlperiode(p, y = year(today()), wahl = FALSE, strict = FALSE)
}
\arguments{
\item{p}{ein String mit dem Namen des Parlaments ("Hessen" - auch in der Form: BLxx mit der Länderkennzahl aus der AGS, etwa "BL06" für den Hessischen Landtag); wenn p numerisch ist, wird es als aw-ID des Parlaments identifiziert (z.B.: 11 für Hessen)}

\item{y}{eine Jahreszahl der gesuchten Wahl bzw. Wahlperiode}

\item{wahl}{Boolean-Variable: TRUE, wenn auch eine Wahl gesucht werden darf, FALSE, wenn nur eine Wahlperiode gefunden werden soll}

\item{strict}{Boolean-Variable: TRUE, wenn nur genaue Treffer akzeptiert werden sollen; in der Default-Einstellung sucht die Funktion nach dem nächsten passenden Wert}
}
\value{
Integer-Wert mit erstem Treffer
}
\description{
`aw_wahlperiode` gibt die ID(s) für eine Wahlperiode zu einem Parlament zurück
}
\details{
Mit aw_wahlperiode kann man sich die IDs einer Wahl bzw. Wahlperiode suchen lassen, und zwar entweder über den Namen (oder die Länderkennzahl) oder die aw-ID, dazu das Jahr (es wird das nächste gültige gesucht). 
Wahlen sind in der aw-Datenbank gewissermaßen besondere Wahlperioden von i.d.R. 2 Monaten Länge. Mit dem Filter Wahl=FALSE kann man dafür sorgen, dass sie nicht gefunden werden; sonst listet die Funktion sie auch auf.
}
\examples{
aw_wahlperiode("Bundestag",2021) liefert 111 (die Bundestags-Wahlperiode 2017-2021)
aw_wahlperiode("Bundestag",2020) liefert 111 (die Bundestags-Wahlperiode 2017-2021)
aw_wahlperiode("Bundestag",2020,wahl=TRUE) liefert 111 (weil die Wahlperiode näher am Zeitpunkt liegt)
aw_wahlperiode("Bundestag",2017,wahl=TRUE) liefert 50 (die Bundestagswahl 2017) - entspricht aw_wahl()
aw_wahl("BL06",2018,wahl=FALSE) liefert (Hessischer Landtag, Wahlperiode 2018-2022)

}
