% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\docType{data}
\name{laender}
\alias{laender}
\title{Liste der Bundesländer}
\format{
A data frame with 17 rows and 4 variables:
\describe{
  \item{id}{die interne aw-ID}
  \item{lkz}{der Länderschlüssel aus der AGS (Bund=0)}
  \item{label}{die Länderbezeichnung}
  \item{llabel_external_long}{die Bezeichnung des Parlaments}
}
}
\source{
\url{http://www.destatis.de/}
}
\usage{
laender
}
\description{
Eine Indexdatei mit der ID, dem Länderschlüssel aus der AGS und dem Namen des Parlaments
}
\keyword{datasets}
