% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\name{aw_get_id}
\alias{aw_get_id}
\title{Request single object by ID}
\usage{
aw_get_id(entity, id, ...)
}
\arguments{
\item{entity}{Entity class, as listed in the factor entities}

\item{id}{}

\item{...}{List of filters}
}
\description{
`aw_get_id` requests data for one single object
}
\details{
This function is used to request one single object, belonging to an entity class,
defined by its ID.
}
\examples{
aw_get_id("politicians",1)

}
