% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\name{aw_get_id}
\alias{aw_get_id}
\title{Datenabfrage}
\usage{
aw_get_id(entity, id, ...)
}
\arguments{
\item{entity}{Entity class, as listed in the factor entities}

\item{id}{}

\item{...}{List of filters}
}
\value{
Dataframe mit allen Datenspalten
}
\description{
`aw_get_id` fragt Daten f체r ein Objekt nach seiner ID ab
}
\details{
Abfrage eines einzelnen Objekts aus der aw-Datenbank. Gibt ein dataframe zur체ck. Der Typ der Entit채t muss bekannt sein und angegeben werden; IDs sind nicht eindeutig: 9188 bezeichnet ein Entit채t vom Typ candidacy-mandate und vom Typ constituency.
}
\examples{
aw_get_id("politicians",1)

}
