% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgeoRdnetenwatchr.R
\name{aw_get_table}
\alias{aw_get_table}
\title{aw_get_table}
\usage{
aw_get_table(entity, ...)
}
\arguments{
\item{entity}{Typ der abgefragten Datenobjekte (Entitäten)}

\item{...}{List of filters}
}
\value{
df mit den Daten (abhängig von der abgefragten Entität)
}
\description{
`aw_get_table` - Datenabfrage in ein Dataframe
}
\details{
Abruf der Daten und Formatierung als Dataframe. Angegeben werden muss immer der Typ der abgefragten Daten - welche Art Entitäten wollen wir aus der Datenbank lesen? - optional sind Filter. Die Filter können dabei entweder in Form eines R-Parameters übergeben werden oder als String. Fragt die API in Blöcken zu 100 Ergebnissen ab.
}
\examples{
aw_get_table("parliaments")
aw_get_table("constituencies",parliament_period=128)
aw_get_table("parliaments","label[cn]=hessen")

}
